[refactor] 프로필 이미지 변경 컴포넌트 분리 및 이미지 비율 유지
